+++ packages/mixins/apollo-query-mixin.js
@@ -116,7 +116,20 @@ export const ApolloQueryMixin = superclass => class extends ApolloElementMixin(s
   }
 
   /**
+   * Exposes the [`ObservableQuery#refetch`](https://www.apollographql.com/docs/react/api/apollo-client.html#ObservableQuery.refetch) method.
+   * @param  {Object} variables The new set of variables. If there are missing variables, the previous values of those variables will be used..
+   * @return {Promise<ApolloQueryResult>}
+   */
+  refetch(...args) {
+    return (
+      this.observableQuery &&
+      this.observableQuery.refetch(...args)
+    );
+  }
+
+  /**
    * Exposes the [`ObservableQuery#setVariables`](https://www.apollographql.com/docs/react/api/apollo-client.html#ObservableQuery.setVariables) method.
+   * @deprecated: This method on ObservableQuery is meant to be private. It will be removed.
    * @param {Object}   variables      The new set of variables. If there are missing variables, the previous values of those variables will be used.
    * @param {boolean=} tryFetch       Try and fetch new results even if the variables haven't changed (we may still just hit the store, but if there's nothing in there will refetch).
    * @param {boolean=} fetchResults   Option to ignore fetching results when updating variables.